+++ src/components/modals/TransferOwnershipModal.jsx
@@ -167,7 +167,7 @@ export function TransferOwnershipModal({
   return (
     <Modal isOpen={isOpen} onOpenChange={onOpenChange}>
       <Modal.Backdrop>
-        <Modal.Container className='p-1' placement='top' scroll='outside'>
+        <Modal.Container className='p-1' placement='center' scroll='outside'>
           <Modal.Dialog className='p-2'>
             <Modal.CloseTrigger className='absolute top-2 right-2' variant='ghost'>
               <IconX />